pyproject.toml
[project]
name = "langchain-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
]
